refactor(config): split resolution module
